refactor: sub-conftest should share the reset_configuration_singleton() method defined in the tests/conftest.py
