+++ tests/config/conftest.py
@@ -11,11 +11,6 @@
 
 import pytest
 
-from taipy.config._config import _Config
-from taipy.config._config_comparator._config_comparator import _ConfigComparator
-from taipy.config._serializer._toml_serializer import _TomlSerializer
-from taipy.config.checker._checker import _Checker
-from taipy.config.checker.issue_collector import IssueCollector
 from taipy.config.config import Config
 from taipy.config.section import Section
 from tests.config.utils.section_for_tests import SectionForTest
@@ -23,22 +18,9 @@ from tests.config.utils.unique_section_for_tests import UniqueSectionForTest
 
 
 @pytest.fixture(scope="function", autouse=True)
-def reset():
+def reset(reset_configuration_singleton):
     reset_configuration_singleton()
     register_test_sections()
-
-
-def reset_configuration_singleton():
-    Config.unblock_update()
-    Config._default_config = _Config()._default_config()
-    Config._python_config = _Config()
-    Config._file_config = _Config()
-    Config._env_file_config = _Config()
-    Config._applied_config = _Config()
-    Config._collector = IssueCollector()
-    Config._serializer = _TomlSerializer()
-    Config._comparator = _ConfigComparator()
-    _Checker._checkers = []
 
 
 def register_test_sections():
